requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic-settings
authlib
httpx
python-multipart
jinja2
python-dotenv
itsdangerous
email-validator
pandas
openpyxl
